backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb
pypdf
sqlalchemy
python-dotenv
requests
